tests/test_cel: Fix CEL unit tests
This is a backport of the test_cel portion of r427870, which was not applied to
the 12 branch. This fixes the compilation issues with the CEL unit tests
introduced by the API changes needed to fix publication of blind transfer
messages.
........

Merged revisions 428104 from file:///srv/subversion/repos/asterisk/branches/12